:truck: Restructure `c_regex_parsing` module files
