+++ woke/woke/d_compile/solc_frontend/solc_runner.py
@@ -5,7 +5,7 @@ import asyncio
 
 from woke.a_config import WokeConfig
 from woke.b_svm import SolcVersionManager
-from woke.c_regex_parsing import SolidityVersion
+from woke.c_regex_parsing.solidity_version import SolidityVersion
 from .input_data_model import (
     SolcInput,
     SolcInputSource,
