+++ woke/woke/x_cli/svm.py
@@ -7,8 +7,7 @@ import click
 
 from woke.a_config import WokeConfig
 from woke.b_svm import SolcVersionManager
-from woke.c_regex_parsing import SolidityVersionExpr
-from woke.c_regex_parsing.a_version import SolidityVersion
+from woke.c_regex_parsing.solidity_version import SolidityVersion, SolidityVersionExpr
 from .console import console
